Changed all Ids to long because of PostGres weirdness

+++ ApplicationLayer/Commands/CreateUserPostCommand.cs
@@ -4,10 +4,10 @@ namespace cms.ApplicationLayer.Commands
 {
     public class CreateUserPostCommand : ICommand<CreateUserPostCommand, CommandResult<Post>>
     {
-        public int UserId;
+        public long UserId;
         public Post Post;
 
-        public CreateUserPostCommand(int userId, Post post)
+        public CreateUserPostCommand(long userId, Post post)
         {
             UserId = userId;
             Post = post;
