+++ ApplicationLayer/Commands/DeleteUsersCommand.cs
@@ -2,11 +2,11 @@ using System.Collections.Generic;
 
 namespace cms.ApplicationLayer.Commands
 {
-    public class DeleteUsersCommand : ICommand<DeleteUsersCommand, CommandResult<List<int>>>
+    public class DeleteUsersCommand : ICommand<DeleteUsersCommand, CommandResult<List<long>>>
     {
-        public IList<int> UserIds;
+        public IList<long> UserIds;
 
-        public DeleteUsersCommand(IList<int> userIds)
+        public DeleteUsersCommand(IList<long> userIds)
         {
             UserIds = userIds;
         }
